// src/common/types/search/search.enum.ts
export enum SearchIndexType {
	Products = 'products',
}

export enum SearchFilterUIType {
  Range = 'range',
  Checkbox = 'checkbox'
}

export enum SearchUploadDataSource {
  Import = 'import',
  Custom = 'custom'
}

export enum SearchIndexDataSource {
  Headder = 'headder',
  Attributes = 'attributes'
}

export enum SearchUISection {
  Main = 'main',
  Attributes = 'attributes',
  Preview = 'preview'
}